www: working stat. component and move network to model
